feat :코드 서비스 관련 예외처리 진행
