+++ src/main/java/goorm/code_challenge/ide/dto/reponse/WrongResponse.java
@@ -1,4 +1,4 @@
-package goorm.code_challenge.ide.dto;
+package goorm.code_challenge.ide.dto.reponse;
 
 import java.util.Map;
