+++ src/main/java/goorm/code_challenge/ide/dto/reuest/CodeSubmission.java
@@ -1,15 +1,18 @@
-package goorm.code_challenge.ide.dto;
+package goorm.code_challenge.ide.dto.reuest;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 
 @Getter
-public class FeedbackRequest {
+public class CodeSubmission {
 	@NotNull(message = "방 번호는 필수 입력값입니다")
 	private Long roomId;
 	@NotNull(message = "문제 번호는 필수 입력값입니다")
 	private Long problemId;
-	@NotNull(message = "작성자 번호는 필수 입력값입니다")
-	private Long userId;
+	@NotBlank(message = "정답 코드는 필수 입력값입니다")
+	private String code;
+	@NotBlank(message = "언어는 필수 입력값입니다")
+	private String compileLanguage;
+
 }
